+++ migrations/mytbdb/sqls/sqlite3/20230120171718-remove-unused-fields-down.sql
@@ -1,0 +1,1 @@
+-- noop - will not restore removed fields